changes in min max

+++ app/src/main/java/pl/lizardproject/qe2015/NumberListParser.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public class NumberListParser {
 
-    public static List<Integer> parseList(String list, String separator) {
+    public static List<Double> parseList(String list, String separator) {
         final String[] split = list.split(separator);
-        List<Integer> numbers = new ArrayList<>(split.length);
-        for(String str : split) {
+        List<Double> numbers = new ArrayList<>(split.length);
+        for (String str : split) {
             try {
-                numbers.add(Integer.parseInt(str));
-            } catch(NumberFormatException ignored) {
+                numbers.add(Double.parseDouble(str));
+            } catch (NumberFormatException ignored) {
             }
         }
         return numbers;
